requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asn1crypto==1.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.1
distro==1.9.0
filelock==3.17.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
numpy==2.2.3
openai==1.62.0
packaging==24.2
platformdirs==4.3.6
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
sniffio==1.3.1
snowflake-connector-python==3.13.2
sortedcontainers==2.4.0
tomlkit==0.13.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
